requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.8.1
django-filter==2.2.0
django-filters==0.2.1
django-search-views==0.3.7
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
